feat(Utilities): split vector and matrix extensions
